// src/src/types/request.d.ts
export interface ITriggerFrequencyPayload {
    frequencyType: string;
    every: number;
    on: string | null;
    starting: string | null;
    selectedValues: string[] | null;
    calendarType: string;
    fiscalValue: string | null;
}